--- foundations/foundation-common/src/main/java/io/servicecomb/foundation/token/AuthenticationTokenManager.java
@@ -1,9 +1,0 @@
-package io.servicecomb.foundation.token;
-
-public interface AuthenticationTokenManager {
-
-	default public String getToken(){return "";}
-	
-	public boolean valid(String token);
-	
-}
