+++ handlers/pom.xml
@@ -34,7 +34,7 @@
     <module>handler-bizkeeper</module>
     <module>handler-flowcontrol-qps</module>
     <module>handler-loadbalance</module>
-    <module>handler-auth</module>
+    <module>handler-publickey-auth</module>
   </modules>
 
 </project>
