+++ samples/auth-sample/auth-consumer/src/main/resources/config/log4j.properties
