+++ samples/auth-sample/auth-provider/src/main/resources/config/log4j.properties
