+++ service-registry/src/main/java/io/servicecomb/serviceregistry/api/response/MicroserviceInstanceResponse.java
@@ -1,0 +1,18 @@
+package io.servicecomb.serviceregistry.api.response;
+
+import io.servicecomb.serviceregistry.api.registry.MicroserviceInstance;
+
+public class MicroserviceInstanceResponse {
+
+	private MicroserviceInstance instance;
+
+	public MicroserviceInstance getInstance() {
+		return instance;
+	}
+
+	public void setInstance(MicroserviceInstance instance) {
+		this.instance = instance;
+	}
+	
+	
+}
